+++ src/es/ladroncorps/AndroidManifest.xml
@@ -2,7 +2,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
     <application>
         <activity
-            android:name=".es.tumanhwas.TheSecretCorpsUrlActivity"
+            android:name=".es.tumanhwas.LadronCorpsUrlActivity"
             android:excludeFromRecents="true"
             android:exported="true"
             android:theme="@android:style/Theme.NoDisplay">
